refactor-move prometheus from main to metrics.go
